pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gsdk"
version = "1.0.1"
authors = [
  { name="ProgVM", email="progvminc@gmail.com" },
]
description = "Fast, lightweight SDK for Google Gemini API."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
]
dependencies = [
    "google-genai>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gsdk*"]

[project.urls]
Homepage = "https://github.com/ProgVM/gsdk"
Repository = "https://github.com/ProgVM/gsdk.git"
Issues = "https://github.com/ProgVM/gsdk/issues"